fatsoma-scraper-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
apscheduler==3.10.4
aiohttp==3.9.1
supabase==2.22.2
python-dotenv==1.2.1
